Revert float64 change

+++ timeseries/src/autogluon/timeseries/models/local/abstract_local_model.py
@@ -140,8 +140,7 @@ class AbstractLocalModel(AbstractTimeSeriesModel):
             data = data.groupby(level=ITEMID, sort=False).tail(self.max_ts_length)
 
         df = pd.DataFrame(data).reset_index(level=ITEMID)
-        # Statistical models lose accuracy if float32 dtype is used
-        all_series = (ts.astype("float64") for _, ts in df.groupby(by=ITEMID, as_index=False, sort=False)[self.target])
+        all_series = (ts for _, ts in df.groupby(by=ITEMID, as_index=False, sort=False)[self.target])
 
         # timeout ensures that no individual job takes longer than time_limit
         # TODO: a job started late may still exceed time_limit - how to prevent that?
